+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 TBD
+
 ### [2.20.8] - 2024-01-09
 
 - Set function `createExpression` as static.