+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-TBD
+- Updated GitHub actions.
 
 ### [2.20.8] - 2024-01-09