+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change log
 
-## [Unreleased]
+### [Unreleased]
+
+### [2.20.9] - 2024-03-20
 
 - Updated CI actions.
 - Completed tests.
 - Added new tests.
+- Updated psalm configuration and annotations.
+- Removed `$trim` parameter for `createExpression` function.
 
 ### [2.20.8] - 2024-01-09
 
@@ -123,6 +127,7 @@ as possible. See [Security Advisory: ZF2014-01](https://framework.zend.com/secur
 - Initial release
 
 [Unreleased]: https://github.com/laurentmuller/HighchartsBundle/compare/1.7...HEAD
+[2.20.9]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.8...2.20.9
 [2.20.8]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.7...2.20.8
 [2.20.7]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.6...2.20.7
 [2.20.6]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.5...2.20.6